Enables product detail view from home and favorites
Allows users to navigate to a product details view from both the home and favorites screens by tapping on a product card.

Adds selectedProduct property to FavoritesViewModel to manage the presentation of the detail view.

Refactors the presentation of the detail view to use a common DetailsView, injecting the necessary dependencies.